More changes! Alsp added Anvil block and BinaryStream
