+++ src/pocketmine/inventory/InventoryType.php
@@ -32,6 +32,9 @@ class InventoryType{
 	const CRAFTING = 4;
 	const WORKBENCH = 5;
 	const STONECUTTER = 6;
+	const BREWING_STAND = 7;
+	const ANVIL = 8;
+	const ENCHANT_TABLE = 9;
 
 	private static $default = [];
 
@@ -59,7 +62,10 @@ class InventoryType{
 		static::$default[static::FURNACE] = new InventoryType(3, "Furnace", 2);
 		static::$default[static::CRAFTING] = new InventoryType(5, "Crafting", 1); //4 CRAFTING slots, 1 RESULT
 		static::$default[static::WORKBENCH] = new InventoryType(10, "Crafting", 1); //9 CRAFTING slots, 1 RESULT
-		static::$default[static::STONECUTTER] = new InventoryType(10, "Crafting", 3); //9 CRAFTING slots, 1 RESULT
+		static::$default[static::STONECUTTER] = new InventoryType(10, "Crafting", 1); //9 CRAFTING slots, 1 RESULT
+		static::$default[static::ENCHANT_TABLE] = new InventoryType(1, "Enchant", 4); //1 INPUT/OUTPUT
+		static::$default[static::BREWING_STAND] = new InventoryType(4, "Brewing", 5); //1 INPUT, 3 POTION
+		static::$default[static::ANVIL] = new InventoryType(3, "Brewing", 6); //2 INPUT, 1 OUTPUT
 	}
 
 	/**
